refactor(data): update import paths and enhance npcQualities table
- Update import paths from '../../types/type' to '../types/type' for consistency
- Add description field to npcQualities table to improve documentation
