+++ src/data/themeTable.ts
@@ -1,4 +1,4 @@
-import { Table } from '../../types/type';
+import { Table } from '../types/type';
 
 const themeTable: Table = {
   name: 'Theme Table',
